feat(agentbundle): move skill-spec lint + pack evals into agentbundle 0.17.0
Ports `tools/lint-skill-spec.py` into `agentbundle catalogue lint --deep`
(new `[lint]` optional extra, PyYAML-backed, CAT-S001–S006 diagnostics) and
`tools/run-pack-evals.py` into `agentbundle pack evals run --catalogue-root`.
Rewires CI to use the package commands. Retires the three standalone tools/.
Fixes Windows backslash in adapt pending report and scope isolation in tests.
Slims AGENTS.local.md; adds packages/AGENTS.md for Python-package rules.
