+++ .github/workflows/pack-evals.yml
@@ -43,13 +43,8 @@ jobs:
         with:
           python-version: "3.11"
 
-      - name: Install agentbundle (projection path) + tool deps
-        run: |
-          pip install -e packages/agentbundle/
-          # Hash-pinned: pack-evals.yml is the repo's first managed-secret workflow;
-          # --require-hashes ensures the sdist/wheel bytes match PyPI's published digests.
-          # Regenerate with: pip-compile --generate-hashes tools/requirements.txt
-          pip install --require-hashes -r tools/requirements-evals-locked.txt
+      - name: Install agentbundle (projection path)
+        run: pip install -e packages/agentbundle/
 
       - name: Install the claude CLI (activation detector)
         # Pinned: the activation event's stream shape is version-sensitive, so
@@ -69,7 +64,7 @@ jobs:
           packs="${PACKS_INPUT:-core converters}"
           for pack in $packs; do
             echo "::group::pack-evals $pack"
-            python tools/run-pack-evals.py --pack "$pack"
+            agentbundle pack evals run --pack "$pack" --catalogue-root .
             echo "::endgroup::"
           done
