+++ packages/agentbundle/pyproject.toml
@@ -4,10 +4,13 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "agentbundle"
-version = "0.16.2"
+version = "0.17.0"
 description = "npm for your coding agent. Install packs of skills, subagents, and hooks into any repo, for every major agent."
 requires-python = ">=3.11"
 dependencies = []
+
+[project.optional-dependencies]
+lint = ["pyyaml>=6.0"]
 authors = [{ name = "eugenelim", email = "eugenelim@users.noreply.github.com" }]
 license = { text = "Apache-2.0 OR MIT" }
 readme = "README.md"
