--- tools/lint-skill-spec.py
@@ -1,768 +1,0 @@
-#!/usr/bin/env python3
-"""Foundation linter for the agentskills.io specification.
-
-Reference: https://agentskills.io/specification
-           https://agentskills.io/skill-creation/evaluating-skills
-
-Walks both the projection (.claude/skills/*/SKILL.md) and the seeds
-(packs/*/.apm/skills/*/SKILL.md) and enforces every mechanical rule in
-the spec. Designed as a contract for skill authors going forward — every
-spec rule has a check even if no skill in the current tree trips it.
-
-Companion to tools/lint-agent-artifacts.py: that linter is the
-multi-artifact hygiene gate (skills + subagents + commands); this one is
-the spec-only deep dive scoped to SKILL.md. The two overlap on the
-top-level frontmatter shape; the duplication is intentional defensive
-depth — the two lints have different lifecycles and the projection /
-seed walk pair belongs here, not there.
-
-Exit codes:
-  0 — every checked skill is clean (warns and infos are non-fatal).
-  1 — at least one skill tripped an error-level rule.
-
-A fixture mode is supported for self-testing:
-  LINT_ROOT=tools/fixtures/<dir> python3 tools/lint-skill-spec.py
-"""
-
-from __future__ import annotations
-
-import json
-import os
-import pathlib
-import re
-import subprocess
-import sys
-import tomllib
-
-# Windows cp1252 guard — reconfigure stdout/stderr to UTF-8 before any print.
-sys.stdout.reconfigure(encoding="utf-8", errors="strict")
-sys.stderr.reconfigure(encoding="utf-8", errors="backslashreplace")
-
-try:
-    import yaml
-except ImportError as exc:  # pragma: no cover — env-setup failure path
-    print(
-        "✖ This linter requires PyYAML. Install with: "
-        "pip install -r tools/requirements.txt "
-        "(or `pip install 'pyyaml>=6.0'` if you're not in the repo root).",
-        file=sys.stderr,
-    )
-    raise SystemExit(2) from exc
-
-
-def _repo_root() -> pathlib.Path:
-    try:
-        result = subprocess.run(
-            ["git", "rev-parse", "--show-toplevel"],
-            capture_output=True, text=True, check=False,
-        )
-        if result.returncode == 0 and result.stdout.strip():
-            return pathlib.Path(result.stdout.strip())
-    except FileNotFoundError:
-        pass
-    return pathlib.Path.cwd()
-
-
-# Per agentskills.io/specification § Frontmatter — the closed top-level
-# key set. Project-specific data lives nested under `metadata:`.
-ALLOWED_SKILL_KEYS = {"name", "description", "license", "compatibility",
-                      "metadata", "allowed-tools"}
-# Per agentskills.io/specification § Frontmatter — `name` must be
-# kebab-case (`^[a-z0-9]+(-[a-z0-9]+)*$`) and 1–64 chars.
-KEBAB = re.compile(r"^[a-z0-9]+(-[a-z0-9]+)*$")
-
-# Per agentskills.io/specification § Directory layout — the four blessed
-# subdirectories. `evals/` adds its own canonical layout (`evals.json`
-# plus `evals/files/<fixture>`) per agentskills.io/skill-creation/evaluating-skills.
-BLESSED_SUBDIRS = {"scripts", "references", "assets", "evals"}
-
-# Not from the spec — project policy. Dev-time directories the linter
-# ignores when checking layout. Scoped to things package managers
-# create: `npm install`, `python -m venv`, Python's bytecode cache.
-# Names like `dist/`, `build/`, `.cache/` are legitimate authored
-# content in some skills, so they intentionally stay warn-eligible.
-IGNORED_DEV_DIRS = {"node_modules", ".venv", "venv", "__pycache__"}
-
-# Body path regexes — the install-prefix rule is absolute per the brief.
-RE_ABS_PATH = re.compile(r"(/Users/|/home/|/opt/|/var/|/etc/|C:\\)")
-# Catches both `.claude/skills/<name>/...` and the bare `.claude/skills/`
-# mention (the brief flags work-loop:412 "any skill in `.claude/skills/`"
-# as a violation — the install path itself is environment-specific and
-# shouldn't be baked into a SKILL.md body, with or without a skill name
-# after it). Same for the seed path.
-RE_INSTALL_PATH = re.compile(
-    r"(?<!~)(?<![\w/])"
-    r"(\.claude/skills/(?:[a-z0-9-]+/)?|"
-    r"packs/[a-z0-9-]+/\.apm/skills/(?:[a-z0-9-]+/)?)"
-)
-# `evals/` follows its own canonical layout (`evals/evals.json`,
-# `evals/files/<fixture>`) per agentskills.io/skill-creation/evaluating-skills,
-# so the depth check is scoped to scripts/references/assets only.
-RE_DEEP_SAME_SKILL = re.compile(
-    r"(?<![\w./~])((?:scripts|references|assets)/[a-z0-9_./-]+/[a-z0-9_.-]+\.[a-z0-9]+)"
-)
-
-# Distinguishes flow-style `allowed-tools: [Read, Grep]` from block-style
-# (subsequent indented `- Read` lines). PyYAML parses both as a Python
-# list — the error message uses this regex against the source frontmatter
-# to keep the block-vs-flow distinction the spec test asserts on.
-RE_ALLOWED_TOOLS_FLOW = re.compile(r"^allowed-tools\s*:\s*\[", re.MULTILINE)
-
-# Scalar shapes acceptable inside `metadata:`. The agentskills.io spec
-# describes metadata as a "mapping of strings"; in practice projects
-# need booleans (e.g. `credentialed: true`), integers, and floats
-# (e.g. `version: 1.0`). PyYAML returns the natural Python types for
-# each. ``None`` covers `key:` with an empty value, which prior versions
-# of this parser normalised to the empty string.
-META_SCALAR_TYPES = (str, bool, int, float, type(None))
-
-# Per YAML 1.2 § 5.3 (Reserved Indicators). At the leading position of an
-# unquoted plain scalar these characters give the scalar a YAML-structural
-# meaning the author almost certainly didn't intend. Wrapping the value
-# in double quotes makes any of them safe. We refuse them at lint time
-# with a policy-named message so the author gets a remediation hint
-# rather than PyYAML's downstream parse error (or worse, a silent
-# anchor-consumption like `&foo bar` → `bar`).
-YAML_LEADING_INDICATORS = "#[]{}|>&*@`%!?"
-
-# Block-scalar markers and their chomping indicators. A `description:`
-# value position holding any of these turns the description into a
-# multi-line scalar, which the cross-IDE norm refuses for portability.
-FOLDED_LITERAL_MARKERS = (">", "|", ">-", "|-", ">+", "|+")
-
-
-class _DuplicateKeyError(Exception):
-    """Raised by the custom YAML mapping constructor on a duplicate key.
-
-    See the matching helper in tools/lint-agent-artifacts.py — defensive
-    depth is intentional, the two lints have different lifecycles.
-    """
-
-    def __init__(self, key: object, line: int) -> None:
-        self.key = key
-        self.line = line
-
-
-class _FrontmatterLoader(yaml.SafeLoader):
-    """SafeLoader subclass that rejects duplicate mapping keys."""
-
-
-def _construct_mapping_no_dups(loader, node, deep=False):
-    if not isinstance(node, yaml.MappingNode):
-        raise yaml.constructor.ConstructorError(
-            None, None,
-            f"expected a mapping node, got {node.id}",
-            node.start_mark,
-        )
-    mapping: dict = {}
-    for key_node, value_node in node.value:
-        key = loader.construct_object(key_node, deep=deep)
-        if key in mapping:
-            raise _DuplicateKeyError(key, key_node.start_mark.line + 1)
-        mapping[key] = loader.construct_object(value_node, deep=deep)
-    return mapping
-
-
-_FrontmatterLoader.add_constructor(
-    yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG,
-    _construct_mapping_no_dups,
-)
-
-
-def _check_description_source(
-    fm_lines: list[str],
-) -> tuple[int, str] | None:
-    """Source-level policy checks on the raw `description:` line.
-
-    Runs before PyYAML parse so each violation surfaces with a
-    policy-named message and remediation, instead of PyYAML's generic
-    `mapping values are not allowed here` (Kiro #8329 colon-space) or
-    silent anchor consumption (`description: &foo bar` -> `bar`).
-    PyYAML stays as the backstop for anything not enumerated here.
-
-    Returns None on clean, or `(line_index, message)` where line_index
-    is 0-based into `fm_lines` (callers convert to the file-absolute
-    line for err()). Returns on the first `description:` line found;
-    later duplicates are caught by the duplicate-key path.
-    """
-    for i, line in enumerate(fm_lines):
-        m = re.match(r"^description:\s*(.*)$", line)
-        if not m:
-            continue
-        value = m.group(1).rstrip()
-        # Folded / literal block markers - agentskills.io requires
-        # single-line scalars; some downstream loaders fail on
-        # multi-line even when YAML parses cleanly.
-        if value in FOLDED_LITERAL_MARKERS:
-            return i, (
-                f"description must be a single-line scalar; folded/literal "
-                f"block syntax ({value!r}) is not portable"
-            )
-        # Continuation lines (next line indented and non-blank) =
-        # multi-line in disguise, whether the first line is empty or
-        # not. `.strip()` filters out trailing-whitespace artifacts so
-        # we only fire on actual indented content.
-        if i + 1 < len(fm_lines):
-            nxt = fm_lines[i + 1]
-            if nxt.strip() and (nxt.startswith(" ") or nxt.startswith("\t")):
-                return i, (
-                    "description must be a single-line scalar; "
-                    "continuation lines (indented next line) are not portable"
-                )
-        if value == "":
-            return None  # genuinely empty - missing-required check downstream
-        # Quoted scalars escape the source-level rules: YAML handles
-        # their content uniformly across parsers.
-        if (
-            len(value) >= 2
-            and value[0] == value[-1]
-            and value[0] in ('"', "'")
-        ):
-            return None
-        # Unquoted scalar - apply leading + mid-scalar rules. Anchor
-        # `&` and alias `*` get a sharper message because PyYAML
-        # silently consumes them (the highest-impact case): the value
-        # mutates rather than the parse failing.
-        leading = value[0]
-        if leading in ("&", "*"):
-            name = "anchor" if leading == "&" else "alias"
-            return i, (
-                f"description starts with YAML {name} indicator {leading!r} "
-                f"in an unquoted scalar; the YAML parser will consume the "
-                f"{name} and the value silently mutates. Wrap in double quotes."
-            )
-        if leading in YAML_LEADING_INDICATORS:
-            return i, (
-                f"description starts with YAML indicator {leading!r} in an "
-                f"unquoted scalar; wrap value in double quotes (the parser "
-                f"may consume the indicator and silently alter the value)"
-            )
-        if ": " in value:
-            return i, (
-                "description contains ': ' in an unquoted scalar; wrap "
-                "value in double quotes (Kiro silently drops skills with "
-                "this pattern from agent discovery -- kirodotdev/Kiro#8329)"
-            )
-        # Comment trap: any whitespace followed by `#` ends a plain
-        # scalar in YAML. `\s#` catches both ` # ` (space-padded) and
-        # ` #x` (no space after hash) - PyYAML truncates both.
-        if re.search(r"\s#", value):
-            return i, (
-                "description contains whitespace-then-'#' in an unquoted "
-                "scalar; the YAML parser treats everything from '#' as a "
-                "comment and truncates the value. Wrap in double quotes."
-            )
-        return None
-    return None
-
-
-def main() -> int:
-    os.chdir(_repo_root())
-    lint_root_env = os.environ.get("LINT_ROOT")
-    root = pathlib.Path(lint_root_env or ".").resolve()
-    if not root.exists():
-        msg = (f"LINT_ROOT={lint_root_env!r} does not exist"
-               if lint_root_env else f"root path {root} does not exist")
-        print(f"✖ {msg}", file=sys.stderr)
-        return 1
-    error_count = 0
-    warn_count = 0
-
-    def relpath(path) -> pathlib.Path:
-        p = pathlib.Path(path)
-        if not p.is_absolute():
-            return p
-        try:
-            return p.resolve().relative_to(root)
-        except ValueError:
-            return p
-        except (OSError, RuntimeError):
-            # `resolve()` raises OSError on Python 3.11 and RuntimeError
-            # on 3.12+ when a symlink loop is in the path. Fall back to
-            # the un-resolved path so the err() message still localises.
-            return p
-
-    def err(path, msg: str, line=None) -> None:
-        nonlocal error_count
-        rel = relpath(path)
-        loc = f"{rel}:{line}" if line else str(rel)
-        print(f"✖ {loc}: {msg}", file=sys.stderr)
-        error_count += 1
-
-    def warn(path, msg: str, line=None) -> None:
-        nonlocal warn_count
-        rel = relpath(path)
-        loc = f"{rel}:{line}" if line else str(rel)
-        print(f"⚠ {loc}: {msg}", file=sys.stderr)
-        warn_count += 1
-
-    def info(path, msg: str) -> None:
-        rel = relpath(path)
-        print(f"ℹ {rel}: {msg}")
-
-    def ok(msg: str) -> None:
-        print(f"✓ {msg}")
-
-    # ── Frontmatter parser ────────────────────────────────────────────────
-    # PyYAML's safe_load via a duplicate-rejecting Loader; mirrored in
-    # tools/lint-agent-artifacts.py. Defensive depth is intentional — the
-    # two lints have different lifecycles. The returned 5-tuple includes
-    # ``fm_text`` (frontmatter source) so downstream checks can do
-    # source-level inspection (e.g. flow-vs-block list distinction).
-    def parse_frontmatter(path: pathlib.Path):
-        """Return (fields, body_start_line, body, error, fm_text)."""
-        # Read bytes first so BOM detection runs before any decode that
-        # would silently strip it. Each BOM has its own policy message.
-        try:
-            raw = path.read_bytes()
-        except OSError as exc:
-            # Match the wording the outer `check_skill` handler used to
-            # emit for the same failure class (symlink loops, vanished
-            # files, permission denials) so tree-E and any downstream
-            # log scrapers see the same substring.
-            return None, 0, "", f"could not read skill: {exc}", ""
-        if raw.startswith(b"\xef\xbb\xbf"):
-            return None, 0, "", (
-                "UTF-8 BOM detected at file start; save the file as UTF-8 "
-                "without BOM (some skill loaders treat the BOM as content "
-                "and fail to find the --- frontmatter fence)"
-            ), ""
-        if raw.startswith(b"\xff\xfe") or raw.startswith(b"\xfe\xff"):
-            return None, 0, "", (
-                "UTF-16 BOM detected at file start; save the file as UTF-8 "
-                "without BOM"
-            ), ""
-        try:
-            text = raw.decode("utf-8")
-        except UnicodeDecodeError as exc:
-            return None, 0, "", f"SKILL.md is not valid UTF-8: {exc}", ""
-        lines = text.splitlines()
-        if not lines or lines[0].strip() != "---":
-            return None, 0, text, None, ""
-        end = None
-        for i in range(1, len(lines)):
-            if lines[i].strip() == "---":
-                end = i
-                break
-        if end is None:
-            return None, 0, text, "frontmatter opened with --- but never closed", ""
-        fm_lines = lines[1:end]
-        fm_text = "\n".join(fm_lines)
-        body_start_line = end + 2
-        body = "\n".join(lines[end + 1:])
-        # Source-level description policy checks run before PyYAML parse
-        # so the author gets a policy-named message + remediation instead
-        # of PyYAML's downstream error (or a silent anchor consumption).
-        # The helper returns (fm_relative_line, message); we convert to
-        # the file-absolute line (frontmatter starts at file line 1; the
-        # fence opens at line 1, so fm_line 0 maps to file line 2).
-        desc_check = _check_description_source(fm_lines)
-        if desc_check is not None:
-            fm_line_idx, desc_msg = desc_check
-            file_line = fm_line_idx + 2
-            return None, 0, text, f"line {file_line}: {desc_msg}", fm_text
-        try:
-            # B506: Loader is a yaml.SafeLoader subclass (class def above).
-            fields = yaml.load(fm_text, Loader=_FrontmatterLoader)  # nosec B506
-        except _DuplicateKeyError as exc:
-            return None, 0, text, (
-                f"duplicate frontmatter key {exc.key!r} (line {exc.line + 1})"
-            ), fm_text
-        except yaml.YAMLError as exc:
-            mark = getattr(exc, "problem_mark", None)
-            problem = getattr(exc, "problem", None) or str(exc)
-            if mark is not None:
-                return None, 0, text, (
-                    f"malformed frontmatter (line {mark.line + 2}): {problem}"
-                ), fm_text
-            return None, 0, text, f"malformed frontmatter: {problem}", fm_text
-        if fields is None:
-            fields = {}
-        if not isinstance(fields, dict):
-            return None, 0, text, (
-                "frontmatter must be a mapping at the top level "
-                f"(got {type(fields).__name__})"
-            ), fm_text
-        return fields, body_start_line, body, None, fm_text
-
-    # ── Per-skill checks ──────────────────────────────────────────────────
-    def check_frontmatter(path: pathlib.Path, fields: dict, fm_text: str) -> None:
-        # Top-level key whitelist — spec § Frontmatter.
-        unknown = sorted(set(fields) - ALLOWED_SKILL_KEYS)
-        if unknown:
-            err(path, f"unknown top-level frontmatter keys: {unknown} "
-                      f"(allowed: {sorted(ALLOWED_SKILL_KEYS)})")
-
-        # name — required, 1–64 chars, kebab regex, matches parent dir.
-        name = fields.get("name")
-        if not isinstance(name, str) or not name:
-            err(path, "missing required key: name")
-        else:
-            if not (1 <= len(name) <= 64):
-                err(path, f"name {name!r} must be 1–64 chars (got {len(name)})")
-            if not KEBAB.match(name):
-                err(path, f"name {name!r} must match "
-                          f"^[a-z0-9]+(-[a-z0-9]+)*$ (kebab-case, no "
-                          f"leading/trailing/double hyphens)")
-            elif name != path.parent.name:
-                err(path, f"name {name!r} does not match directory "
-                          f"{path.parent.name!r}")
-
-        # description — required, non-empty, ≤1024 chars.
-        desc = fields.get("description")
-        if not isinstance(desc, str) or not desc:
-            err(path, "missing required key: description (must be non-empty)")
-        elif len(desc) > 1024:
-            err(path, f"description exceeds 1024 chars (got {len(desc)})")
-
-        # license — if present, non-empty string.
-        if "license" in fields:
-            lic = fields["license"]
-            if not isinstance(lic, str) or not lic:
-                err(path, "'license' must be a non-empty string when present")
-
-        # compatibility — if present, 1–500 chars.
-        if "compatibility" in fields:
-            compat = fields["compatibility"]
-            if not isinstance(compat, str) or not compat:
-                err(path, "'compatibility' must be a non-empty string when present")
-            elif len(compat) > 500:
-                err(path, f"compatibility exceeds 500 chars (got {len(compat)})")
-
-        # metadata — if present, must be a mapping. The agentskills.io
-        # spec describes it as a mapping of strings; in practice the
-        # `metadata:` escape hatch carries booleans (e.g. `credentialed:
-        # true`), version numbers, and lists. PyYAML returns the natural
-        # Python types for each. Project-specific value validation (e.g.
-        # `metadata.credentialed` must be a literal YAML boolean;
-        # `metadata.primitive-class` must be one of the two allowed
-        # strings) is the job of lint-agent-artifacts.py — this linter
-        # enforces only the spec's structural shape so future extensions
-        # don't need a code change here.
-        if "metadata" in fields:
-            meta = fields["metadata"]
-            # `metadata:` with no value parses to ``None``; the rare
-            # ``metadata: ""`` form parses to the empty string. Treat
-            # both as "no project-specific data".
-            if meta == "" or meta is None:
-                pass
-            elif not isinstance(meta, dict):
-                err(path, f"'metadata' must be a nested mapping "
-                          f"(got {type(meta).__name__})")
-            else:
-                for mk, mv in meta.items():
-                    if isinstance(mv, list):
-                        for item in mv:
-                            if not isinstance(item, META_SCALAR_TYPES):
-                                err(path, f"'metadata.{mk}' list entries "
-                                          f"must be scalars (got "
-                                          f"{type(item).__name__})")
-                                break
-                    elif isinstance(mv, dict):
-                        # Doubly-nested mappings are allowed under the
-                        # spec's escape hatch — deeper structure is the
-                        # project's problem, not the spec linter's.
-                        pass
-                    elif not isinstance(mv, META_SCALAR_TYPES):
-                        err(path, f"'metadata.{mk}' must be a scalar or a "
-                                  f"list of scalars (got "
-                                  f"{type(mv).__name__})")
-
-        # allowed-tools — if present, MUST be a space-separated string,
-        # NOT a list. Spec § Frontmatter is explicit on this. PyYAML
-        # parses both block-style (`- Read` indented under the key) and
-        # flow-style (`[Read, Grep]`) as Python lists; we distinguish
-        # the two by scanning the source frontmatter so the error
-        # message faithfully names which list shape the author used.
-        if "allowed-tools" in fields:
-            tools = fields["allowed-tools"]
-            if isinstance(tools, list):
-                if RE_ALLOWED_TOOLS_FLOW.search(fm_text):
-                    err(path, "'allowed-tools' must be a space-separated string, "
-                              "not a YAML flow-style list")
-                else:
-                    shape = ("an empty YAML flow list" if tools == []
-                             else "a YAML block list")
-                    err(path, f"'allowed-tools' must be a space-separated "
-                              f"string, not {shape}")
-            elif not isinstance(tools, str) or not tools:
-                err(path, "'allowed-tools' must be a space-separated string "
-                          "when present")
-
-    def check_body(path: pathlib.Path, body: str, body_start_line: int) -> None:
-        body_lines = body.splitlines()
-        n = len(body_lines)
-        if n > 1000:
-            err(path, f"body exceeds 1000 lines (got {n}); the spec "
-                      f"recommends keeping SKILL.md under 500 lines")
-        elif n > 500:
-            warn(path, f"body exceeds 500 lines (got {n}); the spec "
-                       f"recommends staying under 500")
-
-        for offset, line in enumerate(body_lines):
-            line_no = body_start_line + offset
-            abs_match = RE_ABS_PATH.search(line)
-            if abs_match:
-                err(path, f"absolute system path in body: "
-                          f"{abs_match.group(0)!r}", line=line_no)
-            for install in RE_INSTALL_PATH.finditer(line):
-                hit = install.group(1)
-                err(path, f"install-path reference in body: {hit!r} "
-                          f"— skill bodies must use skill-relative paths for "
-                          f"own files and name-only references for other "
-                          f"skills", line=line_no)
-            # Don't double-warn on deep same-skill paths if the install
-            # match already fired on this line.
-            if not RE_INSTALL_PATH.search(line):
-                for deep in RE_DEEP_SAME_SKILL.finditer(line):
-                    hit = deep.group(1)
-                    warn(path, f"same-skill file reference deeper than one "
-                               f"level: {hit!r} (spec recommends ≤1 level)",
-                         line=line_no)
-
-    def check_layout(skill_dir: pathlib.Path, path: pathlib.Path) -> None:
-        for child in sorted(skill_dir.iterdir()):
-            if child.is_dir():
-                if child.name in IGNORED_DEV_DIRS:
-                    continue
-                if child.name not in BLESSED_SUBDIRS:
-                    warn(path, f"non-blessed top-level subdirectory: "
-                               f"{child.name!r} (spec recommends "
-                               f"{sorted(BLESSED_SUBDIRS)} as the canonical "
-                               f"layout)")
-            else:
-                if child.name == "SKILL.md":
-                    continue
-                info(path, f"loose file at skill root: {child.name!r} "
-                           f"(allowed by spec; logged for visibility)")
-
-    def check_eval_queries(eval_queries_json: pathlib.Path) -> None:
-        """Validate evals/eval_queries.json — the Tier-A trigger-eval
-        convention (pack-activation-evals spec): a flat JSON array whose
-        every element is {query: non-empty str, should_trigger: bool}.
-        Distinct from the output-quality evals.json schema; a skill may
-        ship either file, both, or neither."""
-        try:
-            data = json.loads(eval_queries_json.read_text(encoding="utf-8"))
-        except json.JSONDecodeError as exc:
-            err(eval_queries_json,
-                f"evals/eval_queries.json is not valid JSON: {exc}")
-            return
-        if not isinstance(data, list):
-            err(eval_queries_json,
-                "evals/eval_queries.json must be a JSON array at top level")
-            return
-        for idx, entry in enumerate(data):
-            if not isinstance(entry, dict):
-                err(eval_queries_json,
-                    f"eval_queries[{idx}] must be an object")
-                continue
-            query = entry.get("query")
-            if not isinstance(query, str) or not query:
-                err(eval_queries_json,
-                    f"eval_queries[{idx}].query must be a non-empty string")
-            st = entry.get("should_trigger")
-            # `bool` is a subclass of `int` in Python — require a real
-            # bool so a literal `0`/`1` does not silently pass as the
-            # trigger flag (mirrors the id bool-guard above).
-            if not isinstance(st, bool):
-                err(eval_queries_json,
-                    f"eval_queries[{idx}].should_trigger must be a boolean "
-                    f"(got {type(st).__name__})")
-
-    def check_evals_json(skill_dir: pathlib.Path, evals_json: pathlib.Path,
-                         skill_name: str | None) -> None:
-        try:
-            data = json.loads(evals_json.read_text(encoding="utf-8"))
-        except json.JSONDecodeError as exc:
-            err(evals_json, f"evals/evals.json is not valid JSON: {exc}")
-            return
-        if not isinstance(data, dict):
-            err(evals_json, "evals/evals.json must be a JSON object at top level")
-            return
-        sn = data.get("skill_name")
-        if not isinstance(sn, str) or not sn:
-            err(evals_json, "evals.json 'skill_name' must be a non-empty string")
-        elif skill_name and sn != skill_name:
-            err(evals_json, f"evals.json skill_name {sn!r} does not match "
-                            f"skill name {skill_name!r}")
-        evals_list = data.get("evals")
-        if not isinstance(evals_list, list):
-            err(evals_json, "evals.json 'evals' must be a list")
-            return
-        seen_ids: set = set()
-        for idx, entry in enumerate(evals_list):
-            if not isinstance(entry, dict):
-                err(evals_json, f"evals[{idx}] must be an object")
-                continue
-            eid = entry.get("id")
-            # `bool` is a subclass of `int` in Python — a literal
-            # `"id": true` would otherwise pass the type gate. Reject
-            # explicitly so the type-check matches author intent.
-            if not isinstance(eid, (int, str)) or isinstance(eid, bool):
-                err(evals_json, f"evals[{idx}].id must be int or str "
-                                f"(got {type(eid).__name__})")
-            elif eid in seen_ids:
-                err(evals_json, f"evals[{idx}] duplicate id {eid!r} "
-                                f"— ids must be unique within evals.json")
-            else:
-                seen_ids.add(eid)
-            prompt = entry.get("prompt")
-            if not isinstance(prompt, str) or not prompt:
-                err(evals_json, f"evals[{idx}].prompt must be a non-empty string")
-            expected = entry.get("expected_output")
-            if not isinstance(expected, str) or not expected:
-                err(evals_json, f"evals[{idx}].expected_output must be a "
-                                f"non-empty string")
-            files = entry.get("files")
-            if files is not None:
-                if not isinstance(files, list):
-                    err(evals_json, f"evals[{idx}].files must be a list")
-                else:
-                    skill_dir_resolved = skill_dir.resolve()
-                    for fpath in files:
-                        if not isinstance(fpath, str) or not fpath:
-                            err(evals_json, f"evals[{idx}].files entry must "
-                                            f"be a non-empty string")
-                            continue
-                        resolved = (skill_dir / fpath).resolve()
-                        # Path-traversal guard: an entry like
-                        # `../../some-other-skill/x` must not escape the
-                        # skill dir even if the target happens to exist.
-                        try:
-                            resolved.relative_to(skill_dir_resolved)
-                        except ValueError:
-                            err(evals_json, f"evals[{idx}].files entry "
-                                            f"{fpath!r} resolves outside "
-                                            f"the skill directory "
-                                            f"({resolved})")
-                            continue
-                        if not resolved.exists():
-                            err(evals_json, f"evals[{idx}].files entry "
-                                            f"{fpath!r} does not exist "
-                                            f"(resolved to {resolved})")
-            asserts = entry.get("assertions")
-            if asserts is not None:
-                if not isinstance(asserts, list):
-                    err(evals_json, f"evals[{idx}].assertions must be a list")
-                else:
-                    for ai, a in enumerate(asserts):
-                        if not isinstance(a, str) or not a:
-                            err(evals_json, f"evals[{idx}].assertions[{ai}] "
-                                            f"must be a non-empty string")
-
-    def check_evals(skill_dir: pathlib.Path, path: pathlib.Path,
-                    skill_name: str | None) -> None:
-        """Dispatch over a skill's evals/ directory. A skill may ship the
-        output-quality evals/evals.json (Tier B), the trigger evals/
-        eval_queries.json (Tier A), both, or neither — but an evals/ dir
-        with neither recognised file is an error."""
-        evals_dir = skill_dir / "evals"
-        if not evals_dir.exists() or not evals_dir.is_dir():
-            return
-        evals_json = evals_dir / "evals.json"
-        eval_queries_json = evals_dir / "eval_queries.json"
-        if not evals_json.exists() and not eval_queries_json.exists():
-            err(path, "evals/ directory present but neither evals/evals.json "
-                      "nor evals/eval_queries.json is present")
-            return
-        if evals_json.exists():
-            check_evals_json(skill_dir, evals_json, skill_name)
-        if eval_queries_json.exists():
-            check_eval_queries(eval_queries_json)
-
-    def check_skill(path: pathlib.Path) -> None:
-        fields, body_start, body, ferr, fm_text = parse_frontmatter(path)
-        if ferr:
-            err(path, ferr)
-            return
-        if fields is None:
-            err(path, "missing YAML frontmatter (--- ... ---)")
-            return
-        check_frontmatter(path, fields, fm_text)
-        if not body.strip():
-            err(path, "body is empty")
-        check_body(path, body, body_start)
-        check_layout(path.parent, path)
-        check_evals(path.parent, path, fields.get("name"))
-
-    # ── Walk both roots ───────────────────────────────────────────────────
-    walk_roots: list[pathlib.Path] = []
-    projection = root / ".claude" / "skills"
-    if projection.exists():
-        walk_roots.append(projection)
-    packs_root = root / "packs"
-    if packs_root.exists():
-        walk_roots.extend(sorted(packs_root.glob("*/.apm/skills")))
-
-    skill_count = 0
-    for walk in walk_roots:
-        for skill_md in sorted(walk.glob("*/SKILL.md")):
-            skill_count += 1
-            before_err = error_count
-            before_warn = warn_count
-            try:
-                check_skill(skill_md)
-            except (OSError, RuntimeError) as exc:
-                # Symlink loops, permission denials, vanished files
-                # between glob and check — surface as an error against
-                # the offending skill rather than killing the whole walk.
-                # RuntimeError covers Python 3.12+'s `pathlib.resolve()`
-                # error class for symlink loops (older Pythons raise
-                # OSError for the same condition).
-                err(skill_md, f"could not read skill: {exc}")
-            if error_count == before_err and warn_count == before_warn:
-                ok(f"{relpath(skill_md)}")
-
-    # ── [pack.evals].skills coverage (pack-activation-evals spec, AC7) ─────
-    # A pack-level pass over packs/*/pack.toml: every entry in the optional
-    # `[pack.evals].skills` allowlist must name a real skill directory under
-    # the pack that ships evals/eval_queries.json. A pack with no [pack.evals]
-    # block is a no-op. pack.toml is source-only (never projected), so this
-    # reads packs/<pack>/ directly; the eval_queries.json *schema* is checked
-    # on both the projection and source by the per-skill walk above.
-    if packs_root.exists():
-        for pack_toml in sorted(packs_root.glob("*/pack.toml")):
-            pack_dir = pack_toml.parent
-            try:
-                manifest = tomllib.loads(pack_toml.read_text(encoding="utf-8"))
-            except (OSError, tomllib.TOMLDecodeError) as exc:
-                err(pack_toml, f"could not parse pack.toml: {exc}")
-                continue
-            evals_cfg = manifest.get("pack", {}).get("evals")
-            if evals_cfg is None:
-                continue
-            if not isinstance(evals_cfg, dict):
-                err(pack_toml, "[pack.evals] must be a table")
-                continue
-            skills = evals_cfg.get("skills", [])
-            if not isinstance(skills, list):
-                err(pack_toml, "[pack.evals].skills must be an array of strings")
-                continue
-            for entry in skills:
-                if not isinstance(entry, str) or not entry:
-                    err(pack_toml, f"[pack.evals].skills entry must be a "
-                                   f"non-empty string (got {entry!r})")
-                    continue
-                skill_dir = pack_dir / ".apm" / "skills" / entry
-                if not skill_dir.is_dir():
-                    err(pack_toml, f"[pack.evals].skills names {entry!r} but "
-                                   f"{relpath(skill_dir)} is not a skill directory")
-                    continue
-                if not (skill_dir / "evals" / "eval_queries.json").is_file():
-                    err(pack_toml, f"[pack.evals].skills names {entry!r} but it "
-                                   f"ships no evals/eval_queries.json")
-
-    print()
-    print(f"Skills checked: {skill_count} "
-          f"({error_count} error(s), {warn_count} warning(s)).")
-    if error_count:
-        print()
-        print(f"Skill-spec lint: failed ({error_count} error(s)).")
-        return 1
-    print("Skill-spec lint: passed.")
-    return 0
-
-
-if __name__ == "__main__":
-    sys.exit(main())
