+++ tools/test-all.py
@@ -49,7 +49,9 @@ TESTS: list[tuple[str, list[str]]] = [
     ("lint-agent-artifacts", ["bash", "tools/test-lint-agent-artifacts.sh"]),
     ("lint-catalogue-seeds", [sys.executable, "tools/test-lint-catalogue-seeds.py"]),
     ("lint-knowledge",       ["bash", "tools/test-lint-knowledge.sh"]),
-    ("lint-skill-spec",      [sys.executable, "tools/test-lint-skill-spec.py"]),
+    ("lint-skill-spec",      [sys.executable, "-m", "pytest",
+                              "packages/agentbundle/tests/unit/test_catalogue_skill_spec_lint.py",
+                              "-v"]),
     ("llm-judge-cross-pack-eval", [sys.executable, "tools/test-llm-judge-cross-pack-eval.py"]),
     ("loop-cohort",          ["bash", "tools/test-loop-cohort.sh"]),
     ("pre-pr",               ["bash", "tools/test-pre-pr.sh"]),
